Fix race conditions in watch mode (#911)
This commit fixes the watch mode for several issues, which are best
highlighted by a long application startup:

- there can now only be a single compilation request at a time. Previously,
if compilation took more time than the time it takes to get another watch
change request, 2 compilations could happen concurrently
- application restarts are queued. If an application is requested to be
restarted, we have to wait for the previous process to be killed before
starting the new version. We do this by using a lock. We can safely
discard restart requests if there is already a request in queue, because
it would use the latest version in any case
- the test resources service keep alive file should not use the session
start time, because that time changes when using watch mode. This was
trigerring the shutdown of the server even in shared mode.
- the test resources keep alive file was deleted too eagerly. It should
only be deleted IF we know the server will be shutdown. Previously it
was deleted because a request was made, but in the case of watch mode,
other runs will need the service.

+++ micronaut-maven-plugin/src/main/java/io/micronaut/maven/testresources/TestResourcesHelper.java
@@ -185,15 +185,18 @@ public class TestResourcesHelper {
                 // the build was started, so we put a file to indicate to the stop
                 // mojo that it should not stop the server.
                 Path keepalive = getKeepAliveFile();
-                Files.write(keepalive, "true".getBytes());
-                Runtime.getRuntime().addShutdownHook(new Thread(() -> {
-                    // Make sure that if the build is interrupted, e.g using CTRL+C, the keepalive file is deleted
-                    try {
-                        Files.delete(keepalive);
-                    } catch (IOException e) {
-                        // ignore
-                    }
-                }));
+                // Test is because we may be running in watch mode
+                if (!Files.exists(keepalive)) {
+                    Files.write(keepalive, "true".getBytes());
+                    Runtime.getRuntime().addShutdownHook(new Thread(() -> {
+                        // Make sure that if the build is interrupted, e.g using CTRL+C, the keepalive file is deleted
+                        try {
+                            deleteKeepAliveFile();
+                        } catch (MojoExecutionException e) {
+                            // ignore, we're in a shutdown hook
+                        }
+                    }));
+                }
             }
         }
     }
@@ -270,13 +273,8 @@ public class TestResourcesHelper {
         if (!enabled) {
             return;
         }
-        if (Files.exists(getKeepAliveFile())) {
+        if (isKeepAlive()) {
             log.info("Keeping Micronaut Test Resources service alive");
-            try {
-                Files.delete(getKeepAliveFile());
-            } catch (IOException e) {
-                throw new MojoExecutionException("Failed to delete keepalive file", e);
-            }
             return;
         }
         try {
@@ -292,9 +290,23 @@ public class TestResourcesHelper {
         }
     }
 
-    private void doStop() throws IOException {
-        Path settingsDirectory = getServerSettingsDirectory();
-        ServerUtils.stopServer(settingsDirectory);
+    private void doStop() throws IOException, MojoExecutionException {
+        try {
+            Path settingsDirectory = getServerSettingsDirectory();
+            ServerUtils.stopServer(settingsDirectory);
+        } finally {
+            deleteKeepAliveFile();
+        }
+    }
+
+    private void deleteKeepAliveFile() throws MojoExecutionException {
+        if (Files.exists(getKeepAliveFile())) {
+            try {
+                Files.delete(getKeepAliveFile());
+            } catch (IOException e) {
+                throw new MojoExecutionException("Failed to delete keepalive file", e);
+            }
+        }
     }
 
     private Path getServerSettingsDirectory() {
@@ -306,7 +318,7 @@ public class TestResourcesHelper {
 
     private Path getKeepAliveFile() {
         Path tmpDir = Path.of(System.getProperty("java.io.tmpdir"));
-        return tmpDir.resolve("keepalive-" + mavenSession.getStartTime().getTime());
+        return tmpDir.resolve("keepalive-" + mavenSession.getRequest().getBuilderId());
     }
 
     private Path serverSettingsDirectoryOf(Path buildDir) {
